requirements.txt
streamlit==1.33.0
spacy==3.7.2
pdfminer.six==20221105
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl